requirements.txt
# Computer Vision Anomaly Detection System

# Core Module Dependencies
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# YOLO Dependencies
ultralytics>=8.0.0

# ByteTrack Dependencies
lap>=0.4.0
cython_bbox>=0.1.3

# Audio Processing
pygame>=2.1.0
sounddevice>=0.4.0

# Utilities
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and Logging
pyyaml>=6.0
colorlog>=6.6.0

# Development and Testing
pytest>=6.2.0
pytest-cov>=3.0.0


# numba>=0.56.0  # Uncomment for JIT compilation
# cupy-cuda11x>=10.0.0  # Uncomment for additional GPU acceleration

# Platform-specific dependencies
# Windows OS
pywin32>=227; sys_platform == "win32"

# System Information
psutil>=5.8.0

# Data Handling
pandas>=1.3.0

# Visualization Enhancement
plotly>=5.0.0

# Threading and Async
asyncio>=3.4.3 